CLD-4329 Update options such that textures have an explicit opacity and visibility props
1) It changes `OptionsSize/Color` from inheriting `OptionsOpacity`.
Because i think it just needs the ability of data mapping, so there is a
new type `_BaseOptionsData`(not sure about name). `OptionsOpacity` can
then just inherit that as well.
2) It creates a new type `OptionsStaticOpacity`, for things that dont
have data mapped opacities like textures. and uses this in
`OptionsTexture`
3) It adds `visibility` to `OptionsTexture` with the assumption that
textures are not elements so it just gets its own attributes of opacity
and visibility without inheriting `_BaseElementOptions`